Align command creation names
